Remove unused stuff

+++ server/src/main/java/database/CharacterService.java
@@ -22,34 +22,29 @@ import server.maps.SavedLocationType;
 @Slf4j
 public class CharacterService {
 
-    public static void deleteWhereCharacterId(Connection con, String sql, int id)
-            throws SQLException {
+    public static void deleteWhereCharacterId(Connection con, String sql, int id) throws SQLException {
         PreparedStatement ps = con.prepareStatement(sql);
         ps.setInt(1, id);
         ps.executeUpdate();
         ps.close();
     }
 
-    public static void deleteWhereCharacterName(Connection con, String sql, String name)
-            throws SQLException {
+    public static void deleteWhereCharacterName(Connection con, String sql, String name) throws SQLException {
         PreparedStatement ps = con.prepareStatement(sql);
         ps.setString(1, name);
         ps.executeUpdate();
         ps.close();
     }
 
-    public static void saveAchievement(
-            FinishedAchievements finishedAchievements, int account_id, int characterId) {
+    public static void saveAchievement(FinishedAchievements finishedAchievements, int account_id, int characterId) {
         try (var con = DatabaseConnection.getConnection()) {
             if (finishedAchievements.isChanged()) {
                 var ps = con.prepareStatement("DELETE FROM achievements WHERE accountid = ?");
                 ps.setInt(1, account_id);
                 ps.executeUpdate();
                 ps.close();
-                ps =
-                        con.prepareStatement(
-                                "INSERT INTO achievements(charid, achievementid, accountid)"
-                                        + " VALUES(?, ?, ?)");
+                ps = con.prepareStatement(
+                        "INSERT INTO achievements(charid, achievementid, accountid)" + " VALUES(?, ?, ?)");
                 for (Integer achievementid : finishedAchievements.getFinishedAchievements()) {
                     ps.setInt(1, characterId);
                     ps.setInt(2, achievementid);
@@ -65,15 +60,12 @@ public class CharacterService {
 
     public static void saveSkillMacro(SavedSkillMacro savedSkillMacro, int characterId) {
         try (var con = DatabaseConnection.getConnection()) {
-            deleteWhereCharacterId(
-                    con, "DELETE FROM skillmacros WHERE characterid = ?", characterId);
+            deleteWhereCharacterId(con, "DELETE FROM skillmacros WHERE characterid = ?", characterId);
             for (int i = 0; i < 5; i++) {
                 final SkillMacro macro = savedSkillMacro.getSkillMacros()[i];
                 if (macro != null) {
-                    var ps =
-                            con.prepareStatement(
-                                    "INSERT INTO skillmacros (characterid, skill1, skill2, skill3,"
-                                        + " name, shout, position) VALUES (?, ?, ?, ?, ?, ?, ?)");
+                    var ps = con.prepareStatement("INSERT INTO skillmacros (characterid, skill1, skill2, skill3,"
+                            + " name, shout, position) VALUES (?, ?, ?, ?, ?, ?, ?)");
                     ps.setInt(1, characterId);
                     ps.setInt(2, macro.getSkill1());
                     ps.setInt(3, macro.getSkill2());
@@ -120,8 +112,7 @@ public class CharacterService {
     public static void unban(int accId) {
         try (var con = DatabaseConnection.getConnection()) {
             PreparedStatement ps =
-                    con.prepareStatement(
-                            "UPDATE accounts SET banned = 0 and banreason = '' WHERE id = ?");
+                    con.prepareStatement("UPDATE accounts SET banned = 0 and banreason = '' WHERE id = ?");
             ps.setInt(1, accId);
             ps.executeUpdate();
             ps.close();
@@ -132,10 +123,8 @@ public class CharacterService {
 
     public static int deleteCharacter(final int cid, int accId) {
         try (var con = DatabaseConnection.getConnection()) {
-            PreparedStatement ps =
-                    con.prepareStatement(
-                            "SELECT guildid, guildrank, name FROM characters WHERE id = ? AND"
-                                    + " accountid = ?");
+            PreparedStatement ps = con.prepareStatement(
+                    "SELECT guildid, guildrank, name FROM characters WHERE id = ? AND" + " accountid = ?");
             ps.setInt(1, cid);
             ps.setInt(2, accId);
             ResultSet rs = ps.executeQuery();
@@ -154,54 +143,32 @@ public class CharacterService {
             }
 
             final String name = rs.getString("name");
-            CharacterService.deleteWhereCharacterName(
-                    con, "DELETE FROM `notes` WHERE `to` = ?", name);
-            CharacterService.deleteWhereCharacterName(
-                    con, "DELETE FROM `notes` WHERE `from` = ?", name);
+            CharacterService.deleteWhereCharacterName(con, "DELETE FROM `notes` WHERE `to` = ?", name);
+            CharacterService.deleteWhereCharacterName(con, "DELETE FROM `notes` WHERE `from` = ?", name);
 
             rs.close();
             ps.close();
 
-            CharacterService.deleteWhereCharacterId(
-                    con, "DELETE FROM characters WHERE id = ?", cid);
-            CharacterService.deleteWhereCharacterId(
-                    con, "DELETE FROM monsterbook WHERE charid = ?", cid);
-            CharacterService.deleteWhereCharacterId(
-                    con, "DELETE FROM hiredmerch WHERE characterid = ?", cid);
+            CharacterService.deleteWhereCharacterId(con, "DELETE FROM characters WHERE id = ?", cid);
+            CharacterService.deleteWhereCharacterId(con, "DELETE FROM monsterbook WHERE charid = ?", cid);
+            CharacterService.deleteWhereCharacterId(con, "DELETE FROM hiredmerch WHERE characterid = ?", cid);
             // CharacterService.deleteWhereCharacterId(con, "DELETE FROM cheatlog WHERE characterid
             // = ?", cid);
-            CharacterService.deleteWhereCharacterId(
-                    con, "DELETE FROM mountdata WHERE characterid = ?", cid);
-            CharacterService.deleteWhereCharacterId(
-                    con, "DELETE FROM inventoryitems WHERE characterid = ?", cid);
-            CharacterService.deleteWhereCharacterId(
-                    con, "DELETE FROM famelog WHERE characterid = ?", cid);
-            CharacterService.deleteWhereCharacterId(
-                    con, "DELETE FROM famelog WHERE characterid_to = ?", cid);
-            CharacterService.deleteWhereCharacterId(
-                    con, "DELETE FROM wishlist WHERE characterid = ?", cid);
-            CharacterService.deleteWhereCharacterId(
-                    con, "DELETE FROM buddyentries WHERE owner = ?", cid);
-            CharacterService.deleteWhereCharacterId(
-                    con, "DELETE FROM buddyentries WHERE buddyid = ?", cid);
-            CharacterService.deleteWhereCharacterId(
-                    con, "DELETE FROM keymap WHERE characterid = ?", cid);
-            CharacterService.deleteWhereCharacterId(
-                    con, "DELETE FROM savedlocations WHERE characterid = ?", cid);
-            CharacterService.deleteWhereCharacterId(
-                    con, "DELETE FROM skills WHERE characterid = ?", cid);
-            CharacterService.deleteWhereCharacterId(
-                    con, "DELETE FROM mountdata WHERE characterid = ?", cid);
-            CharacterService.deleteWhereCharacterId(
-                    con, "DELETE FROM skillmacros WHERE characterid = ?", cid);
-            CharacterService.deleteWhereCharacterId(
-                    con, "DELETE FROM trocklocations WHERE characterid = ?", cid);
-            CharacterService.deleteWhereCharacterId(
-                    con, "DELETE FROM queststatus WHERE characterid = ?", cid);
-            CharacterService.deleteWhereCharacterId(
-                    con, "DELETE FROM customqueststatus WHERE characterid = ?", cid);
-            CharacterService.deleteWhereCharacterId(
-                    con, "DELETE FROM inventoryslot WHERE characterid = ?", cid);
+            CharacterService.deleteWhereCharacterId(con, "DELETE FROM mountdata WHERE characterid = ?", cid);
+            CharacterService.deleteWhereCharacterId(con, "DELETE FROM inventoryitems WHERE characterid = ?", cid);
+            CharacterService.deleteWhereCharacterId(con, "DELETE FROM famelog WHERE characterid = ?", cid);
+            CharacterService.deleteWhereCharacterId(con, "DELETE FROM wishlist WHERE characterid = ?", cid);
+            CharacterService.deleteWhereCharacterId(con, "DELETE FROM buddyentries WHERE owner = ?", cid);
+            CharacterService.deleteWhereCharacterId(con, "DELETE FROM buddyentries WHERE buddyid = ?", cid);
+            CharacterService.deleteWhereCharacterId(con, "DELETE FROM keymap WHERE characterid = ?", cid);
+            CharacterService.deleteWhereCharacterId(con, "DELETE FROM savedlocations WHERE characterid = ?", cid);
+            CharacterService.deleteWhereCharacterId(con, "DELETE FROM skills WHERE characterid = ?", cid);
+            CharacterService.deleteWhereCharacterId(con, "DELETE FROM mountdata WHERE characterid = ?", cid);
+            CharacterService.deleteWhereCharacterId(con, "DELETE FROM skillmacros WHERE characterid = ?", cid);
+            CharacterService.deleteWhereCharacterId(con, "DELETE FROM trocklocations WHERE characterid = ?", cid);
+            CharacterService.deleteWhereCharacterId(con, "DELETE FROM queststatus WHERE characterid = ?", cid);
+            CharacterService.deleteWhereCharacterId(con, "DELETE FROM customqueststatus WHERE characterid = ?", cid);
+            CharacterService.deleteWhereCharacterId(con, "DELETE FROM inventoryslot WHERE characterid = ?", cid);
             return 0;
         } catch (Exception e) {
             log.info("Log_Packet_Except.rtf", e);
@@ -214,8 +181,7 @@ public class CharacterService {
         List<CharNameAndId> chars = new LinkedList<>();
         try (var con = DatabaseConnection.getConnection()) {
             PreparedStatement ps =
-                    con.prepareStatement(
-                            "SELECT id, name FROM characters WHERE accountid = ? AND world = ?");
+                    con.prepareStatement("SELECT id, name FROM characters WHERE accountid = ? AND world = ?");
             ps.setInt(1, accountId);
             ps.setInt(2, serverId);
 
@@ -231,8 +197,7 @@ public class CharacterService {
         return chars;
     }
 
-    public static List<MapleCharacter> loadCharacters(
-            MapleClient client, final int serverId, int accountId) {
+    public static List<MapleCharacter> loadCharacters(MapleClient client, final int serverId, int accountId) {
         final List<MapleCharacter> chars = new LinkedList<>();
         final List<Integer> allowedChar = new LinkedList<>();
 
@@ -252,10 +217,8 @@ public class CharacterService {
         if (savedLocations.isChanged()) {
             try (var con = DatabaseConnection.getConnection()) {
                 deleteWhereCharacterId(con, "DELETE FROM savedlocations WHERE characterid = ?", id);
-                var ps =
-                        con.prepareStatement(
-                                "INSERT INTO savedlocations (characterid, `locationtype`, `map`)"
-                                        + " VALUES (?, ?, ?)");
+                var ps = con.prepareStatement(
+                        "INSERT INTO savedlocations (characterid, `locationtype`, `map`)" + " VALUES (?, ?, ?)");
                 ps.setInt(1, id);
                 for (final SavedLocationType savedLocationType : SavedLocationType.values()) {
                     if (savedLocations.getSavedLocation(savedLocationType.getValue()) != -1) {
